feat: useRawSignMessage hook
